// frontend/src/types/voyager.ts
export default interface Voyager {
  _id: string;
  timestamp: string;
  gender: string;
  countryCode: string;
  timezone: string;
  goal: string;
  goalOther: string;
  source: string;
  sourceOther: string;
  countryName: string;
  soloProjectTier: string;
  roleType: string;
  voyageRole: string;
  voyageSignups: string;
  voyageTier: string;
  lat: string;
  lng: string;
  createdAt: string;
  updatedAt: string;
}